Trailer is now an interface
